+++ ExpensesTracker/Presentation/Main/AddTransactionViewController.swift
@@ -1,0 +1,18 @@
+//
+//  AddTransactionViewController.swift
+//  ExpensesTracker
+//
+//  Created by Dmytro Hetman on 06.07.2024.
+//
+
+import UIKit
+
+final class AddTransactionView: UIView {
+    
+}
+
+final class AddTransactionViewController: BaseBackgroundUIViewController {
+    
+    let addTransactionView = AddTransactionView()
+    
+}
