+++ ExpensesTracker/Presentation/Main/ViewModel/MainPageViewModel.swift
@@ -12,47 +12,25 @@ enum NetworkError: Error {
     case decodingFailed
 }
 
-final class ExchangeRateModel {
-    
-    func getBtcUsdRate(completion: @escaping (Result<String, Error>) -> Void) {
-        do {
-            try Network<CurrentBitcoinPriceEndpoint>(APIHost.coindesk)
-                .perform(.get, .currentBitcoinPrice) { result in
-                    
-                switch result {
-                    case .data(let data):
-                        guard let data,
-                              let rates = try? JSONDecoder().decode(BitcoinPriceData.self, from: data)
-                        else {
-                            completion(.failure(NetworkError.decodingFailed))
-                            return
-                        }
-                            
-                        let usdRate = rates.bpi.USD.rate
 
-                        completion(.success(usdRate))
-                    case .error(let error):
-                        completion(.failure(error))
-                    }
-                }
-            
-        } catch {
-            print(error)
-            completion(.failure(error))
-        }
-    }
-    
-}
 
 final class MainPageViewModel {
     
     private let exchangeRateModel: ExchangeRateModel
     
-    var bitcoinUsdRate: Double = 0
+    var bitcoinBalance: Double = 0
+    var transactions: [Transaction] = []
     
     init(exchangeRateModel: ExchangeRateModel) {
         self.exchangeRateModel = exchangeRateModel
         
+    }
+    
+    func topUpBalance(on amount: String, completion: (String) -> Void) {
+        if let topUpValue = Double(amount.replacingOccurrences(of: ",", with: ".")) {
+            bitcoinBalance += topUpValue
+            completion("\(bitcoinBalance)")
+        }
     }
     
     func getBtcUsdRate(completion: @escaping (String) -> Void) {
